+++ src/utils.ts
@@ -1,6 +1,7 @@
 import { invalidValue } from "./errors";
 import { ScAddr } from "./ScAddr";
 import { ScConstruction } from "./ScConstruction";
+import { SnakeToCamelCase } from "./types";
 
 export const transformEdgeInfo = (
   construction: ScConstruction,
@@ -24,3 +25,22 @@ export const transformEdgeInfo = (
     value: aliasIndex,
   };
 };
+
+export const shiftMap = (map: Map<any, any>, to = 1) => {
+  if (to < 1) return;
+
+  let isDone = false;
+  let ind = 0;
+  const mapIterator = map.keys();
+  while (ind < to && !isDone) {
+    const elem = mapIterator.next();
+    isDone = !!elem.done;
+    map.delete(elem.value);
+    ind++;
+  }
+};
+
+export const snakeToCamelCase = <Str extends string>(
+  str: Str
+): SnakeToCamelCase<Str> =>
+  str.replace(/_(\w)/g, (_, p1) => p1.toUpperCase()) as SnakeToCamelCase<Str>;